lou-media-stack/ai-services/recommendation-engine/requirements.txt
flask==3.0.0
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.0
requests==2.31.0
gunicorn==21.2.0
joblib==1.3.0
python-dateutil==2.8.2
scipy==1.11.0
redis==4.6.0
psycopg2-binary==2.9.7
